backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
opencv-python
ffmpeg-python
numpy
scipy
scikit-learn
pandas
aiofiles
tqdm
